test: consolidate per-task model control tests
Merge test_model_routing, test_budget_fallbacks, test_finding_verification,
and test_dedupe_model into a single test_model_control module grouped by
feature area. Also add the runtime.max_context_images field to the runner
prompt test's settings mock so it works with the current runner.
