+++ tests/test_runner_root_prompt.py
@@ -45,7 +45,8 @@ def _patch_engine_scaffold(
             model="openai/gpt-4o",
             reasoning_effort="high",
             force_required_tool_choice=False,
-        )
+        ),
+        runtime=types.SimpleNamespace(max_context_images=None),
     )
     monkeypatch.setattr(runner, "load_settings", lambda: settings)
     monkeypatch.setattr(runner, "configure_sdk_model_defaults", lambda _settings: None)
